requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain-community==0.0.10
pinecone-client==3.0.0
google-generativeai==0.3.2
tiktoken==0.5.2